Add config option to allow *not* saving .orig configs
When admins choose to allow slackpkg new-config to overwrite
currently installed config files with the .new counterparts
from an upgraded package, slackpkg has traditionally created
backups of the config files first, with the backups being
written out with .orig extensions. This commit does not change
the default behavior, but it does allow it to be turned off.

TLDR: set "ORIG_BACKUPS" to "off" in /etc/slackpkg/slackpkg.conf
      if you do not wish to have .orig files left over

Signed-off-by: Robby Workman <rworkman@slackware.com>