Fix (hopefully) blacklisting of specific packages
In short, it seems impossible to blacklist some packages without
catching other undesirable packages. As an example, blacklisting
glibc will also catch glibc-zoneinfo (and the other glibc-*)
packages. It would be good to have the ability to blacklist only
the glibc package without it catching the others.

With this change, if sysadmin wants the blacklist to be "greedy,"
then adding "glibc.*" to the blacklist will do that.

Reported-by: Peter Hyman <pete@peterhyman.com>
Signed-off-by: Robby Workman <rworkman@slackware.com>